app.js: chaining callbacks together
This commit links from weather.getWeather() and geocode.geocodeAddress, making
the app outputs temperatures based on ZIP code and location name inputs, making it
more real-life like.

This commit also abstracts the darksky and Google's geocode API keys. They are stored
in a local folder that will need to be recreated in the client's system if ever there
is one.

NOTE: The app will not work beyond this point without the API keys in place. So the user
must consider that and set it up accordingly withh his own keys.

Ref.: Lecture 35

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>