keys.js: untrack file
This commit removes keys.js from the tracking pointer, for the API keys
for Google Geolocation and DarksSky need to meu used. As these keys are
not to be exposed, I will add them locally but add the file to .gitignore,
so it will not be committed.

This commit completes this action, along with .gitignore insertion.

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>